make coverity-scan happy
* coverity-scan complaints about the fallthrough parts of the switch-statement
  in the base64-encoder. i replaced it with if-statements.

* coverity-scan complaints about pointer arithmetik in md5_final(). i applied
  an explicit typecast to "char*" when calling byte_zero(). this seems to be
  sufficient.

also: several other, mostly cosmetic changes.

+++ platform_msvc.c
@@ -8,16 +8,17 @@
 \*------------------------------------------------------------------*/
 
 #include "platform.h"
+#include <stdint.h>
 
 #define WIN32LEAN_AND_MEAN
 #include <windows.h>
 #include <io.h>
 
-int posix_write(int fd, const void* buf, unsigned int n) {
+int posix_write(int fd, const void* buf, size_t n) {
     return _write(fd, buf, n);
 }
 
-int posix_read(int fd, void* buf, unsigned int n) {
+int posix_read(int fd, void* buf, size_t n) {
     return _read(fd, buf, n);
 }
